fix lint warnings
