+++ packages/web/src/components/ui/button/button.tsx
@@ -1,0 +1,36 @@
+import * as React from 'react';
+import { Slot } from '@radix-ui/react-slot';
+import { type VariantProps } from 'class-variance-authority';
+
+import { cn } from '@/lib/utils';
+import { Loader } from '../loader';
+import { buttonVariants } from './buttonVariants';
+
+export interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
+    VariantProps<typeof buttonVariants> {
+  asChild?: boolean;
+  isLoading?: boolean;
+}
+
+const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
+  ({ className, variant, size, asChild = false, isLoading = false, children, ...props }, ref) => {
+    const Comp = asChild ? Slot : 'button';
+    return (
+      <Comp
+        className={cn(
+          buttonVariants({ variant, size, className }),
+          isLoading && 'relative cursor-not-allowed opacity-70',
+        )}
+        disabled={isLoading || props.disabled}
+        ref={ref}
+        {...props}
+      >
+        {isLoading ? <Loader size="sm" /> : children}
+      </Comp>
+    );
+  },
+);
+Button.displayName = 'Button';
+
+export { Button };
