+++ packages/web/src/components/ui/button/buttonVariants.ts
@@ -1,11 +1,6 @@
-import * as React from 'react';
-import { Slot } from '@radix-ui/react-slot';
-import { cva, type VariantProps } from 'class-variance-authority';
+import { cva } from 'class-variance-authority';
 
-import { cn } from '@/lib/utils';
-import { Loader } from './loader';
-
-const buttonVariants = cva(
+export const buttonVariants = cva(
   'inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0',
   {
     variants: {
@@ -31,32 +26,3 @@ const buttonVariants = cva(
     },
   },
 );
-
-export interface ButtonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
-    VariantProps<typeof buttonVariants> {
-  asChild?: boolean;
-  isLoading?: boolean;
-}
-
-const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
-  ({ className, variant, size, asChild = false, isLoading = false, children, ...props }, ref) => {
-    const Comp = asChild ? Slot : 'button';
-    return (
-      <Comp
-        className={cn(
-          buttonVariants({ variant, size, className }),
-          isLoading && 'relative cursor-not-allowed opacity-70',
-        )}
-        disabled={isLoading || props.disabled}
-        ref={ref}
-        {...props}
-      >
-        {isLoading ? <Loader size="sm" /> : children}
-      </Comp>
-    );
-  },
-);
-Button.displayName = 'Button';
-
-export { Button, buttonVariants };
