+++ packages/web/src/components/ui/button/index.ts
@@ -1,0 +1,2 @@
+export * from './button';
+export * from './buttonVariants';
